Split crypt.rs into several submodules
